requirements.txt
python-dotenv
langchain
langchain_core
langchain-google-genai
langchain-ollama
gradio